db types changes, api calls, date formatting

+++ src/routes/api/[task]/+server.ts
@@ -1,4 +1,4 @@
-import { json } from '@sveltejs/kit';
+import { json, type Cookies } from '@sveltejs/kit';
 import { TOKEN } from '$env/static/private';
 import { db } from '$lib/server/db/index.js';
 import { term } from '$lib/server/db/schema.js';
@@ -6,16 +6,8 @@ import type { Term } from '$lib/types.js';
 import { eq } from 'drizzle-orm';
 
 export const GET = async ({ cookies, params }) => {
-	// authentication
-	const isAuthenticated = cookies.get('admin_auth_obscure');
-	if (isAuthenticated !== TOKEN) {
-		return json({ authenticated: false }, { status: 401 });
-	}
-
-	// If the token is valid, proceed to fetch sensitive data from the database
 	const { task } = params;
 	switch (task) {
-		// terms
 		case 'load-terms':
 			return await loadTerms();
 		// invalid task
@@ -25,37 +17,36 @@ export const GET = async ({ cookies, params }) => {
 };
 
 export const POST = async ({ request, cookies, params }) => {
-	// authentication
-	const isAuthenticated = cookies.get('admin_auth_obscure');
-	if (isAuthenticated !== TOKEN) {
-		return json({ authenticated: false }, { status: 401 });
-	}
-
-	// If the token is valid
 	const data = await request.json();
 	const { task } = params;
 	switch (task) {
-		// terms
 		case 'add-term':
-			return await addTerm(data);
+			return await addTerm(data, cookies);
 		case 'delete-term':
-			return await deleteTerm(data.id);
+			return await deleteTerm(data.id, cookies);
 		// invalid task
 		default:
 			return json({ message: 'Invalid task' }, { status: 400 });
 	}
 };
 
+// tasks
 async function loadTerms() {
 	try {
-		const result = await db.select().from(term);
+		const result = await db.select().from(term).orderBy(term.date, term.timeStart);
 		return json(result); // Return the result as JSON
 	} catch (err) {
 		return json({ message: 'Error fetching data' }, { status: 500 });
 	}
 }
 
-async function addTerm(data: Term) {
+async function addTerm(data: Term, cookies: Cookies) {
+	// authentication
+	const isAuthenticated = cookies.get('admin_auth_obscure');
+	if (isAuthenticated !== TOKEN) {
+		return json({ authenticated: false }, { status: 401 });
+	}
+	// If the token is valid, proceed to fetch sensitive data from the database
 	try {
 		const result = await db.insert(term).values(data);
 		return json(result);
@@ -64,7 +55,13 @@ async function addTerm(data: Term) {
 	}
 }
 
-async function deleteTerm(id: number) {
+async function deleteTerm(id: number, cookies: Cookies) {
+	// authentication
+	const isAuthenticated = cookies.get('admin_auth_obscure');
+	if (isAuthenticated !== TOKEN) {
+		return json({ authenticated: false }, { status: 401 });
+	}
+	// If the token is valid, proceed to fetch sensitive data from the database
 	try {
 		const result = await db.delete(term).where(eq(term.id, id));
 		return json(result);
